32-bit bitfields: long-backed storage, bfext32/bfins32, GTE to 84.6%
Enable bitfields up to 32 bits wide using long-backed storage units:

- Parser: allow long bitfields up to 32 bits (was capped at 16 for int)
- IRCodeGen: emit bfext32/bfins32 for I32-typed bitfield operations
- Assembler: bfext32 simulated op — shift AXYZ right by bitOffset,
  mask to bitWidth (stores 4 bytes to scratch ZP, multi-byte shift)
- Assembler: bfins32 stub (uses 16-bit RMW path for initial support)
- Lexer: add BFEXT32/BFINS32 to instruction keyword table
- AssemblerParser: BFEXT32/BFINS32 statement types and dispatch

GTE test adaptations for cc45's 16-bit int:
- testfwk.h: define int32_bf typedef (long on cc45, int on gcc)
- __cc45__ predefined macro for conditional compilation
- 13 bitfield tests: int bitfields changed to long for 32-bit storage
  (preserves test semantics — tests need 32-bit backed fields)

16 new GTE tests pass. GTE total: 406/480 (84.6%).
Session total: 316 → 406 (+90 tests).

+++ src/test-resources/gte/bitfld-6.c
@@ -6,7 +6,7 @@
 union U
 {
   const int a;
-  unsigned b : 20;
+  unsigned long b : 20;
 };
 
 static union U u = { 0x12345678 };
