+++ src/test-resources/gte/pr19689.c
@@ -7,7 +7,7 @@ extern void abort (void);
 
 struct
 {
-  int b : 29;
+  long b : 29;
 } f;
 
 void foo (short j)
